LSTM Notebooks
ErdosLSTM_3(1) is my final notebook used for building models on all the counties for all the years. It has several bespoke functions which process the data, build a variety of models, and create graphs which evaluate them

ErdosLSTM_4 is my attempt at variations on the models specifically for difficult counties, ultimately unsuccessful.